Migrate away from Telegram Bot Utils and integrate all components into Telegram SDK
