+++ src/features/chat/attachments_describer.py
@@ -9,7 +9,6 @@ from db.schema.tools_cache import ToolsCache, ToolsCacheSave
 from di.di import DI
 from features.audio.audio_transcriber import AudioTranscriber
 from features.chat.supported_files import KNOWN_AUDIO_FORMATS, KNOWN_DOCS_FORMATS, KNOWN_IMAGE_FORMATS
-from features.chat.telegram.sdk.telegram_bot_sdk_utils import TelegramBotSDKUtils
 from features.documents.document_search import DocumentSearch
 from features.images.computer_vision_analyzer import ComputerVisionAnalyzer
 from util import log
@@ -61,7 +60,7 @@ class AttachmentsDescriber:
                     log.d(f"Malformed LLM Input Error: Attachment '{attachment_id}' not found in DB. You may retry only once!"),
                 )
             attachments.append(ChatMessageAttachment.model_validate(attachment_db))
-        self.__attachments = TelegramBotSDKUtils.refresh_attachment_instances(self.__di, attachments)
+        self.__attachments = self.__di.telegram_bot_sdk.refresh_attachment_instances(attachments)
 
     @property
     def __resolution_status(self) -> Result:
